control-atlas-saas/backend/requirements.txt
fastapi
uvicorn
python-multipart
biopython
numpy
PyYAML
google-generativeai
fpdf2
requests
